requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
face-recognition
face-recognition-models
numpy
setuptools
gunicorn
python-dotenv
